chore: move config files in separate directory
