+++ src/cartridges/int_adyen_SFRA/cartridge/adyen/analytics/job/analyticsJob.js
@@ -136,60 +136,6 @@ function sendAnalyticsEvents(requestObjectList) {
   });
 }
 
-function processConfigurationTime() {
-  const query = 'custom.processingStatus = {0}';
-  const queryArgs = [analyticsConstants.processingStatus.NOT_PROCESSED];
-  let customObjectIterator = null;
-  try {
-    // Query the custom objects by processing status
-    customObjectIterator = CustomObjectMgr.queryCustomObjects(
-      analyticsConstants.configurationTimeEventObjectId,
-      query,
-      null,
-      queryArgs,
-    );
-    const analyticsRequest = createAnalyticsRequest(customObjectIterator);
-    sendAnalyticsEvents(analyticsRequest);
-  } catch (e) {
-    AdyenLogs.error_log('Error querying custom objects:', e);
-    throw e;
-  } finally {
-    if (customObjectIterator) {
-      customObjectIterator.close();
-    }
-  }
-}
-
-function clearConfigurationTime() {
-  let customObjectIterator = null;
-  try {
-    const query =
-      'custom.processingStatus = {0} OR custom.processingStatus = {1}';
-    const queryArgs = [
-      analyticsConstants.processingStatus.SKIPPED,
-      analyticsConstants.processingStatus.PROCESSED,
-    ];
-    customObjectIterator = CustomObjectMgr.queryCustomObjects(
-      analyticsConstants.configurationTimeEventObjectId,
-      query,
-      null,
-      queryArgs,
-    );
-    while (customObjectIterator.hasNext()) {
-      const customObject = customObjectIterator.next();
-      Transaction.wrap(() => {
-        CustomObjectMgr.remove(customObject);
-      });
-    }
-  } catch (e) {
-    AdyenLogs.error_log('Error deleting custom object:', e);
-  } finally {
-    if (customObjectIterator) {
-      customObjectIterator.close();
-    }
-  }
-}
-
 function processAnalytics() {
   const query = 'custom.processingStatus = {0}';
   const queryArgs = [analyticsConstants.processingStatus.NOT_PROCESSED];
@@ -246,7 +192,7 @@ function clearAnalytics() {
 
 module.exports = {
   processAnalytics,
-  processConfigurationTime,
   clearAnalytics,
-  clearConfigurationTime,
+  createAnalyticsRequest,
+  sendAnalyticsEvents,
 };
